// packages/mpl-core-kit-lib/src/generated/types/baseLinkedDataKey.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  type Address,
  type Codec,
  combineCodec,
  type Decoder,
  type Encoder,
  type GetDiscriminatedUnionVariant,
  type GetDiscriminatedUnionVariantContent,
  getAddressDecoder,
  getAddressEncoder,
  getDiscriminatedUnionDecoder,
  getDiscriminatedUnionEncoder,
  getStructDecoder,
  getStructEncoder,
  getTupleDecoder,
  getTupleEncoder,
} from '@solana/kit'

import {
  type BasePluginAuthority,
  type BasePluginAuthorityArgs,
  getBasePluginAuthorityDecoder,
  getBasePluginAuthorityEncoder,
} from './index.ts'

export type BaseLinkedDataKey =
  | { __kind: 'LinkedAppData'; fields: readonly [BasePluginAuthority] }
  | { __kind: 'LinkedLifecycleHook'; fields: readonly [Address] }

export type BaseLinkedDataKeyArgs =
  | { __kind: 'LinkedAppData'; fields: readonly [BasePluginAuthorityArgs] }
  | { __kind: 'LinkedLifecycleHook'; fields: readonly [Address] }

// Data Enum Helpers.
export function baseLinkedDataKey(
  kind: 'LinkedLifecycleHook',
  data: GetDiscriminatedUnionVariantContent<BaseLinkedDataKeyArgs, '__kind', 'LinkedLifecycleHook'>['fields'],
): GetDiscriminatedUnionVariant<BaseLinkedDataKeyArgs, '__kind', 'LinkedLifecycleHook'>
export function baseLinkedDataKey(
  kind: 'LinkedAppData',
  data: GetDiscriminatedUnionVariantContent<BaseLinkedDataKeyArgs, '__kind', 'LinkedAppData'>['fields'],
): GetDiscriminatedUnionVariant<BaseLinkedDataKeyArgs, '__kind', 'LinkedAppData'>
export function baseLinkedDataKey<K extends BaseLinkedDataKeyArgs['__kind'], Data>(kind: K, data?: Data) {
  return Array.isArray(data) ? { __kind: kind, fields: data } : { __kind: kind, ...(data ?? {}) }
}

export function getBaseLinkedDataKeyCodec(): Codec<BaseLinkedDataKeyArgs, BaseLinkedDataKey> {
  return combineCodec(getBaseLinkedDataKeyEncoder(), getBaseLinkedDataKeyDecoder())
}
export function getBaseLinkedDataKeyDecoder(): Decoder<BaseLinkedDataKey> {
  return getDiscriminatedUnionDecoder([
    ['LinkedLifecycleHook', getStructDecoder([['fields', getTupleDecoder([getAddressDecoder()])]])],
    ['LinkedAppData', getStructDecoder([['fields', getTupleDecoder([getBasePluginAuthorityDecoder()])]])],
  ])
}
export function getBaseLinkedDataKeyEncoder(): Encoder<BaseLinkedDataKeyArgs> {
  return getDiscriminatedUnionEncoder([
    ['LinkedLifecycleHook', getStructEncoder([['fields', getTupleEncoder([getAddressEncoder()])]])],
    ['LinkedAppData', getStructEncoder([['fields', getTupleEncoder([getBasePluginAuthorityEncoder()])]])],
  ])
}

export function isBaseLinkedDataKey<K extends BaseLinkedDataKey['__kind']>(
  kind: K,
  value: BaseLinkedDataKey,
): value is { __kind: K } & BaseLinkedDataKey {
  return value.__kind === kind
}